chore(deps): consolidate Dependabot updates
TypeScript 7 requires Next's CLI checker, which enforces generated route
export contracts. Keep route modules limited to supported exports so the
grouped dependency upgrade remains buildable.
