+++ .github/workflows/deploy-staging.yml
@@ -21,7 +21,7 @@ jobs:
     environment: staging
     steps:
       - name: Check out repository
-        uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
+        uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7.0.1
 
       - name: Deploy Staging
         id: deployment
