+++ .github/workflows/refresh-preview.yml
@@ -24,7 +24,7 @@ jobs:
     environment: preview
     steps:
       - name: Check out trusted deployment action
-        uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
+        uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7.0.1
         with:
           ref: dev
 
@@ -35,7 +35,7 @@ jobs:
         run: git check-ref-format --branch "$GIT_BRANCH"
 
       - name: Check out target branch
-        uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
+        uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7.0.1
         with:
           ref: refs/heads/${{ inputs.git_branch }}
           path: target
