+++ package.json
@@ -49,16 +49,16 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@ai-sdk/gateway": "^4.0.26",
-    "@aws-sdk/client-s3": "^3.1092.0",
+    "@ai-sdk/gateway": "^4.0.28",
+    "@aws-sdk/client-s3": "^3.1095.0",
     "@base-ui/react": "^1.6.0",
-    "@clerk/nextjs": "^7.5.22",
+    "@clerk/nextjs": "^7.6.1",
     "@js-temporal/polyfill": "^0.5.1",
     "@upstash/ratelimit": "^2.0.8",
     "@upstash/redis": "^1.38.0",
     "@vercel/analytics": "^2.0.1",
-    "@vercel/functions": "^3.7.5",
-    "ai": "^7.0.34",
+    "@vercel/functions": "^3.7.6",
+    "ai": "^7.0.37",
     "beautiful-mermaid": "^1.1.3",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
@@ -71,7 +71,7 @@
     "github-slugger": "^2.0.0",
     "gray-matter": "^4.0.3",
     "heic-decode": "^2.1.0",
-    "lucide-react": "^1.25.0",
+    "lucide-react": "^1.27.0",
     "motion": "^12.42.2",
     "next": "16.3.0-preview.9",
     "next-mdx-remote": "^6.0.0",
@@ -79,12 +79,12 @@
     "react": "^19.2.8",
     "react-day-picker": "^10.0.1",
     "react-dom": "^19.2.8",
-    "rehype-pretty-code": "^0.14.4",
+    "rehype-pretty-code": "^0.14.5",
     "rehype-slug": "^6.0.0",
     "remark-gfm": "^4.0.1",
     "rss": "^1.2.2",
     "server-only": "^0.0.1",
-    "shaders": "3.0.442",
+    "shaders": "3.0.444",
     "sharp": "^0.35.3",
     "shiki": "^4.3.1",
     "subset-font": "^2.5.0",
@@ -95,20 +95,20 @@
   "devDependencies": {
     "@electric-sql/pglite": "^0.5.4",
     "@next/playwright": "16.3.0-preview.9",
-    "@playwright/test": "1.61.1",
+    "@playwright/test": "1.62.0",
     "@tailwindcss/postcss": "^4.3.3",
     "@testing-library/react": "^16.3.2",
     "@types/mdx": "^2.0.14",
-    "@types/node": "^24.13.3",
+    "@types/node": "^26.1.1",
     "@types/pg": "^8.20.0",
     "@types/react": "^19.2.17",
     "@types/react-dom": "^19.2.3",
     "@types/rss": "^0.0.32",
     "drizzle-kit": "^0.31.10",
     "jsdom": "^29.1.1",
-    "shadcn": "^4.14.0",
+    "shadcn": "^4.15.0",
     "tailwindcss": "^4.3.3",
-    "typescript": "^6.0.3",
+    "typescript": "^7.0.2",
     "vitest": "^4.1.10",
     "yaml": "^2.9.0"
   }
